Move child logic into separate module
